the summary of two queue problems
